dev-requirements.txt
djangorestframework==3.11.0
wheel==0.35.1
twine==3.2.0
